Add separate ServiceEntryId span processor
