+++ solarwinds_apm/trace/__init__.py
@@ -1,3 +1,4 @@
 from .forceflush_processor import ForceFlushSpanProcessor
+from .serviceentry_processor import ServiceEntryIdSpanProcessor
 
-__all__ = ["ForceFlushSpanProcessor"]
+__all__ = ["ForceFlushSpanProcessor", "ServiceEntryIdSpanProcessor"]
